requirements.txt
google-genai>=1.0.0
pyyaml>=6.0
rich>=13.0.0
prompt-toolkit>=3.0.0
